Add debugging info to startup log
There are two ways of setting a configuration for this app: config files, and environment variables.

Currently, only the NODE_ENV variable and server url are being printed out on startup.

This adds the database config and the cron config to the details being displayed. This may help debug configuration issues, database issues, and generally making easier to understand what config is being used for example when running in a container.

The app was also defaulting to display "development" as the NODE_ENV when no environment has been set, I've removed this for consistency with `pa11y-dashboard` where we will display `undefined` when this is not set.