Display task id before each line of output
This change makes the task id available inside the log function, so each line of output can be displayed next to the id of the task it belongs to.

This solves the problem of not being able to tell to which task a line of output belongs to, as tasks run and output to the screen concurrently.