Move mongo events from db class to client
Listening for events on the db class of the mongo client is now deprecated. The modern behaviour is to listen to events on the client class.

This PR fixes this behaviour and removes the deprecation warnings when starting webservice (and dashboard)
> (node:20551) DeprecationWarning: Listening to events on the Db class has been deprecated and will be removed in the next major version.

Additionally, the event listeners were set inside the connect() callback, which means that there was the potential for some of the events to be lost. We are now setting up the events before calling connect(), so we can be sure that we will capture all of them.