Validate all ids before they are sent to the DB
Until now, ids passed in a request were not validated by webservice, but were instead sent directly to the database layer for retrieval, relying on the database code to return an appropriate error that would then cause webservice to return a 404.

This was the root cause for all the numerous "ObjectID generation failed" errors happening on pa11y dashboard.

This commit changes this behaviour so all requests made to webservice containing database ids will have those ids validated using joi before they are used in database queries. All queries received with an id in an invalid will now return an "HTTP 400 Bad Request" instead of an "HTTP 404 Not Found", which is why this is potentially a breaking change.

Tests have also been updated accordingly.